+++ packages/lib/src/model/prop.ts
@@ -9,9 +9,19 @@ import { IsOptionalValue } from "../utils/types"
  */
 export interface ModelPropOptions {
   /**
-   * Set to `true` to automatically wrap the property setter in a model action (defaults to `false`).
+   * Set to `true` to automatically generate a property setter in a model action (defaults to `false`).
    */
   readonly setterAction?: boolean
+}
+
+/**
+ * Model property options with setterAction set to true.
+ */
+export interface ModelPropOptionsWithSetterAction extends ModelPropOptions {
+  /**
+   * Set to `true` to automatically generate a property setter in a model action (defaults to `false`).
+   */
+  readonly setterAction: true
 }
 
 /**
@@ -27,13 +37,15 @@ export interface ModelProp<
   TPropCreationValue,
   TIsOptional,
   TInstanceValue = TPropValue,
-  TInstanceCreationValue = TPropCreationValue
+  TInstanceCreationValue = TPropCreationValue,
+  THasSetterAction = never
 > {
   $propValueType: TPropValue
   $propCreationValueType: TPropCreationValue
   $instanceValueType: TInstanceValue
   $instanceCreationValueType: TInstanceCreationValue
   $isOptional: TIsOptional
+  $hasSetterAction: THasSetterAction
 
   defaultFn: (() => TPropValue) | typeof noDefaultValue
   defaultValue: TPropValue | typeof noDefaultValue
@@ -43,10 +55,15 @@ export interface ModelProp<
 }
 
 /**
+ * Any model property.
+ */
+export type AnyModelProp = ModelProp<any, any, any, any, any, any>
+
+/**
  * Model properties.
  */
 export interface ModelProps {
-  [k: string]: ModelProp<any, any, any>
+  [k: string]: AnyModelProp
 }
 
 export type OptionalModelProps<MP extends ModelProps> = {
@@ -73,7 +90,11 @@ export type ModelPropsToInstanceCreationData<MP extends ModelProps> = O.Optional
     [k in keyof MP]: MP[k]["$instanceCreationValueType"]
   },
   OptionalModelProps<MP>
->
+  >
+
+export type ModelPropsToSetterActions<MP extends ModelProps> = {
+  [k in (keyof MP & string) as MP[k]["$hasSetterAction"] & `set${Capitalize<k>}`]: (value: MP[k]["$instanceValueType"]) => void
+}
 
 /**
  * @ignore
@@ -92,6 +113,18 @@ export type MaybeOptionalModelProp<TPropValue, TInstanceValue = TPropValue> = Mo
 >
 
 /**
+ * A model prop that maybe / maybe not is optional, depending on if the value can take undefined, with a setter action.
+ */
+export type MaybeOptionalModelPropWithSetterAction<TPropValue, TInstanceValue = TPropValue> = ModelProp<
+  TPropValue,
+  TPropValue,
+  IsOptionalValue<TPropValue, string, never>,
+  TInstanceValue,
+  TInstanceValue,
+  string
+>
+
+/**
  * A model prop that is definitely optional.
  */
 export type OptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelProp<
@@ -101,6 +134,38 @@ export type OptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelPr
   TInstanceValue,
   TInstanceValue | null | undefined
 >
+
+/**
+ * A model prop that is definitely optional, with a setter action.
+ */
+export type OptionalModelPropWithSetterAction<TPropValue, TInstanceValue = TPropValue> = ModelProp<
+  TPropValue,
+  TPropValue | null | undefined,
+  string,
+  TInstanceValue,
+  TInstanceValue | null | undefined,
+  string
+>
+
+/**
+ * Defines a model property, with an optional function to generate a default value
+ * if the input snapshot / model creation data is `null` or `undefined`.
+ *
+ * Example:
+ * ```ts
+ * x: prop(() => 10) // an optional number, with a default value of 10
+ * x: prop<number[]>(() => []) // an optional number array, with a default empty array
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param defaultFn Default value generator function.
+ * @param options Model property options.
+ * @returns
+ */
+export function prop<TValue>(
+  defaultFn: () => TValue,
+  options: ModelPropOptionsWithSetterAction
+): OptionalModelPropWithSetterAction<TValue>
 
 /**
  * Defines a model property, with an optional function to generate a default value
@@ -140,6 +205,27 @@ export function prop<TValue>(
  */
 export function prop<TValue>(
   defaultValue: OnlyPrimitives<TValue>,
+  options: ModelPropOptionsWithSetterAction
+): OptionalModelPropWithSetterAction<TValue>
+
+/**
+ * Defines a model property, with an optional default value
+ * if the input snapshot / model creation data is `null` or `undefined`.
+ * You should only use this with primitive values and never with object values
+ * (array, model, object, etc).
+ *
+ * Example:
+ * ```ts
+ * x: prop(10) // an optional number, with a default value of 10
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param defaultValue Default primitive value.
+ * @param options Model property options.
+ * @returns
+ */
+export function prop<TValue>(
+  defaultValue: OnlyPrimitives<TValue>,
   options?: ModelPropOptions
 ): OptionalModelProp<TValue>
 
@@ -156,9 +242,25 @@ export function prop<TValue>(
  * @param options Model property options.
  * @returns
  */
+export function prop<TValue>(options: ModelPropOptionsWithSetterAction): MaybeOptionalModelPropWithSetterAction<TValue>
+
+/**
+ * Defines a model property with no default value.
+ *
+ * Example:
+ * ```ts
+ * x: prop<number>() // a required number
+ * x: prop<number | undefined>() // an optional number, which defaults to undefined
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param options Model property options.
+ * @returns
+ */
 export function prop<TValue>(options?: ModelPropOptions): MaybeOptionalModelProp<TValue>
 
-export function prop<TValue>(arg1?: any, arg2?: any): ModelProp<TValue, any, any> {
+// base
+export function prop<TValue>(arg1?: any, arg2?: any): ModelProp<TValue, any, any, TValue, any, any> {
   let def: any
   let opts: ModelPropOptions = {}
   let hasDefaultValue = false
@@ -188,6 +290,7 @@ export function prop<TValue>(arg1?: any, arg2?: any): ModelProp<TValue, any, any
     $isOptional: null as any,
     $instanceValueType: null as any,
     $instanceCreationValueType: null as any,
+    $hasSetterAction: null as any,
 
     defaultFn: hasDefaultValue && isDefFn ? def : noDefaultValue,
     defaultValue: hasDefaultValue && !isDefFn ? def : noDefaultValue,