+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+- Requires TypeScript 4.1.0
+- [BREAKING CHANGE] Class model `setterAction: true` prop option will now generate a setter like `setX` instead.
+- [BREAKING CHANGE] Functional model `setterActions` now just take the field names rather than an object with the setter mapping.
 - `types.enum` no longer requires the generic type in typescript, the enum object as parameter is enough.
 
 ## 0.46.0